requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==4.4.2
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
httplib2==0.21.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.6.2
moviepy==1.0.3
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pip==22.3
proglog==0.1.10
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
